Do nothing when replacing a wiki menu item with itsself

+++ app/controllers/wiki_menu_items_controller.rb
@@ -86,7 +86,7 @@ class WikiMenuItemsController < ApplicationController
   def replace_main_menu_item
     current_page = WikiPage.find params[:id]
 
-    if (current_menu_item = current_page.menu_item) && (page = WikiPage.find_by_id(params[:wiki_page][:id]))
+    if (current_menu_item = current_page.menu_item) && (page = WikiPage.find_by_id(params[:wiki_page][:id])) && current_menu_item != page.menu_item
       create_main_menu_item_for_wiki_page(page, current_menu_item.options)
       current_menu_item.destroy
     end
